brain-python/requirements.txt
fastapi
httpx
pytest
python-dotenv
ts3
uvicorn[standard]
